requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-SQLAlchemy==3.0.2
Flask-Login==0.5.0
Flask-WTF==1.0.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Bootstrap==3.3.7.1
SQLAlchemy==2.0.38
WTForms==3.2.1
email-validator==1.1.3
python-dotenv==0.19.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
Pillow==11.1.0
requests==2.26.0
bcrypt==3.2.0
python-dateutil==2.8.2
Jinja2==3.1.5
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
alembic==1.14.1
blinker==1.9.0
dnspython==2.1.0
idna==3.2
six==1.16.0
visitor==0.1.3
dominate==2.9.1
greenlet==3.1.1
typing-extensions==4.12.2
colorama==0.4.6
Mako==1.3.9 